+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - fix test
 - fix func return in deco
 - fix fmt indent
+- move `Parse` module to `From` module
+- move `highlight` to `Into` module
+
 
 ## [0.7.0]
 - fix long hint cover the screen